Revert "Add code to support Wayland"
This reverts commit 33f80ef17a4564fd6af4fed01de0dd0f3f1e0e2c.

This attempt was completely botched and can't even compile, and even if
it did you would have to choose at compile-time whether to allow X11
*or* Wayland to be used (compiling both and choosing the appropriate one
at runtime is impossible) which might be OK when building for yourself
but is unacceptable when distributing the app to others.

On top of that, rebasing on top of this commit makes git cry, so
reverting it first before doing it right.